Modify token database info handling logic
Currently nanoappTokenDatabaseInfo is only sent if a token database is
found in the binary and the nanoapp binary stored on the host side is
only freed after receiving the corresponding token database info. This
logic should be changed so that if a nanoapp doesn't support tokenized
logging, it should also receive a token database info that suggests and
token database is not found and free the stored binary.

Bug: 242760291
Test: load nanoapp with tokenized logging enabled
Change-Id: I44cf11a5ba04d0496f3479a8ed3e6def14aee583