TransactionManager: Optimize locking
This CL optimizes locking in TransactionManager by
reducing the need to lock for mTimerHandle changes
as well as when deferring the processing of
transactions.

This CL also states that each of the user-provided
callbacks cannot call into TransactionManager methods.
Since we allow the user's start callback to change the
transaction data, we must hold the internal mutex lock
during this callback. To prevent deadlock, we specify
that all callbacks cannot call TransactionManager
methods.

Bug: 333569046
Test: Run chre_unit_tests for TransactionManager
Change-Id: If7fddbeef17ba3ffa574422b8b53ab067b121d89