TransactionManagerTest: Reduce locking
This CL reduces locking in the TransactionManager
tests to reduce flaking by reducing the amount of
time a lock is held to the minimum.

This CL also increases the test from 1 ms retries
to 10 ms retries with a 100 ms timeout.

This CL also changes all waits with no timeout
to waits with a timeout to not block the test during
a flake.

Bug: 330330474
Change-Id: Id55aa34d78c27fa90fdda7cca945d4af32408155
Flag: TEST_ONLY
Test: Run chre_unit_tests for TransactionManager
Test: 1000 iter, all pass