chre: Adjust EPOCH used for CHRE_API_VERSION to 2023
Patch version was using epoch of 2017 Jan 1 which has just rolled over.

Comment says patch version is scoped to the API version/ API version.
Since API was last changed on January 4th, 2024 9:23 AM (to
CHRE_API_VERSION_1_10), it should be safe to move to newer EPOCH. This
CL uses 2023 just to be safe.

Bug: 348994533
Change-Id: I825ec83ccc438a1d8fbc087fcdccd203dcaf8fd7