Fix duplicate detection in CHRE
This CL fixes the duplicate detection logic in CHRE
to ensure that any duplicate message is correctly
flagged. This fixes an issue where multiple duplicates
over a small amount of time may process as non-duplicate
messages.

This CL also fixes an issue with an uninitialized variable.

Bug: 333567700
Change-Id: I4cdc96e8ae86d2a11b6b3e42b71da70c551e3d18
Flag: EXEMPT CHRE change
Test: Run the reliable message test
Test: Run chre_unit_tests