ChreApiTest nanoapp: Reduce stack usage
This CL reduces stack usage in the ChreApiTest
nanoapp by using a statically-allocated
chre_rpc_GeneralEventsMessage. This significantly
reduces the stack space needed when processing these
messages. Since CHRE is single-threaded, this global
will only be accessed by the one CHRE thread at a time
and is safe to use as-is.

Bug: 356663618
Change-Id: I09a1880d065a7830c91d348dcfa78abaca0c9a7f
Flag: TEST_ONLY
Test: Build the nanoapp with -Wframe-larger-than=600
Test: Run the CHRE concurrency test