Adjust wifi timeout to cover entire scan lifetime
Previously the timeout timer was cancelled during handleScanResponse,
and not after delivery of the potential pending ScanEvents that follow.
This led to some missed timeouts in bug cases where follow-up ScanEvents
were promised, but never delievered.

Move the timeout cancellation to the completion of the ScanEvent, in
cases where they are pending.

Bug: 354045837
Test: ./load_aoc.sh
Change-Id: I7a4caf88b10ed730e483ecb1c170c632a421a276