Fix CHPP timeout retry reset recovery logic
When FW service is slow to respond to handshake or requests,
CHPP client will retry TX and/or RESET to re-establish communication.
The timer hanlder would mistakenly thinking it's already in reset
mode and not carry out the actual RESET.

Also increase the default TX retry timeout to match the current
value used in AOC.

b/371080468 to track the task of creating unit tests.

Bug: 354457323
Bug: 368810711
Test: on device
Change-Id: I0df63811d3cc387a5fb23f4e3e346a2e907c52b6