chre/core: Enable a consistent build graph
Instead of conditionally including different headers which rely on
translation units being provided based on the CHRE variant
configuration, strip out the contents of the conditionally enabled
features to ensure that the same core build rule can always be used
regardless of the variant configuration.

Without this change this is impossible because these translation
units which are disabled would get compiled but they would fail to
work because the event loop manager is missing removed methods.

Bug: 376488392
Change-Id: I522387330ce51b703c2d02be8e2f019fdeecc38a