chpp: Don't re-send payload except on retry
Earlier code would always include pending payload in an outbound packet,
which could manifest into an immediate retry that consumes one of the TX
attempts prior to giving up. This would happen any time a TX packet was
waiting on an ACK, and we received any packet other than an ACK for that
pending packet (e.g. duplicate ACK for prior packet, a payload-bearing
packet, or an invalid packet generating a NACK).

Fix this by being more judicious about when we enqueue a TX packet (e.g.
don't do this for a duplicate ACK), and only append payload when we
haven't tried sending it yet or we are explicitly retrying.

Bug: 377999167
Test: FakeLinkSyncTests.OutboundThrottling x 100, presubmit
Change-Id: I72bb749f4dde9056f87054f6387381afd1a9481b