subterra-hal: Adds MessageHubManager class
MessageHubManager tracks the state of all MessageRouter hubs, endpoints,
and sessions, with mappings to do lookups based on HAL client pid, hub
id, endpoint id, and session id.

Each HAL client gets a MessageHubManager::HostHub instance which can be
accessed through lookup APIs in MessageHubManager. A HostHub is created
on the first lookup on that client's pid and is destroyed when the
client disconnects. The HostHub instance will be used from IContextHub
APIs to register the client callback, (un)register endpoints, and
open/close sessions. It is also used to validate messages sent in either
direction over a session and to access the callback to send messages
back to the HAL client.

Later CLs flesh out the implementation and integrate it into the phone
and watch ContextHub HALs.

Bug: 378545373
Test: Builds
Change-Id: Ic92629f2967b31e63b461092224bd447e594a331