subterra-hal: Use single global session id map
Creates a MessageHubManager::Session struct storing weak references to
the host hub, host endpoint, and embedded endpoint, as well as a flag
set when a session open succeeds. On a session open request, a Session
object is created and mapped to the session id.

The weak references in Session removes the need for additional mappings
to session state. Endpoints and hubs can be removed without updating
session state. When necessary (i.e. messages being sent over
a session), the references can be checked to ensure that endpoints are
still valid. If not, session state can be pruned at that time.

Adds MessageHubManager::HostHub::checkSession() and
MessageHubManager::checkSessionAndGetHostHub() for use on host- and
CHRE-originating messages / session open confirmations respectively.

Bug: 378545373
Test: Builds
Change-Id: I5fb4ec907eb577109740ea17272aa4a1f930390a