subterra-hal: Go back to per-hub session maps
Each host hub will be represented with its own MessageHub instance on
the AOC side. As such, each session-related message to the HAL from AOC
will be able to include the hub id. We can make the session state
specific to each HostHub which simplifies the verification logic in a
few places.

Adds a host_hub_id field to the session-related flatbuffer messages.

Bug: 378545373
Test: Builds
Change-Id: Ia56e9b59950329acd7b84624006d4d94290d2c55