Add WWAN CrossValidation detailed comparison
Updates WWAN Xval to compare the contents of the cell_info results.
Initial support is for Nr, LTE, and GSM (5G/4G/3G), though proto
framework is in place for all supported types. Comparisons are limited
to fields which are guaranteed to be populated for cell identity.
Strength comparisons are made only on the primary strength indicator,
and with a margin of error to accommodate for small changes between the
host and CHRE requests.

Additionally updates the test to force a refresh of cell info available
to the host using requestCellInfoUpdate.

Bug: 383769260
Bug: 382533059
Test: wwanCrossValidationTest for preferred network 5G, 4G, and 3G
Change-Id: I2fbdd1fa7e2057e75f3587cddf7b6b48af256206