Reroute heapFreeMessageCallback away from container support lib
The container support library should only be used in instances where all
allocations and frees are done through the library (e.g. in container
utils), and code needs to be shared between the framework and nanoapps.
In this case, we have a nanoapp-specific util and an allocation that is
done separately, so it should use chreHeapFree().

Fixes: 385345503
Test: presubmit
Change-Id: I8449ca359e06a8155accf5363b94bbbf7b457908