Handles recovery for unexpected CHPP open failures
In some scenarios, the app layer can get stuck in a bad state where an open request fails but client expects the link to be in a working state (for example, if a re-open command fails). Since it's possible that this condition is transient, this CL adds logic to reset the transport layer (and thus triggering an open-retry at the app layer) when a new request comes in the closed state.

Bug: 384725643
Test: Reset remote endpoint and inject link-disablement, confirm re-open command can succeed when link is re-enabled
Change-Id: Ia17aba4f569ecde28f5d2aaea01bafa91947afca