Clear outgoing packet code on transport perm failure state
Leaving the packet code to its old value causes a problem where if the app-layer request times out, the transport will think it needs to send an empty packet to the remote endpoint. Since the link has been paused, we clear the packet code to avoid additional transmissions until next recovery is attempted.

Bug: 384725643
Test: Tests pass
Change-Id: I03d62c7e04f76729011ac6f0d83700d3f43f0657