contexthub: Change host message hub init logic
Changes the ContextHub HAL <-> HostMessageHubManager logic for
initializing the embedded hub state in the HAL and the host hub proxies
in CHRE. Instead of having a single reset message which sends all host
state to CHRE and a single response with all of the embedded state, use
the existing hub and endpoint registration messages. This avoids the
need for several copies and large allocations CHRE-side during
initialization.

Bug: 390447515
Test: HAL builds. CHRE-side simulation tests pass.
Change-Id: I5bacbb6eac5c0d532e5a295415946396ad90ee7a