Fix potential race conditions in chppTransportSendReset
Add appropriate mutex protection to protect against race conditions where multiple threads are attempting to reset the transport link. This could happen, for example, if the RX thread is sending a RESET-ACK, and the TX thread is sending a RESET, simultaneously. Without the mutex, the assertion in chppTransportSendReset could be triggerred, since there's no guarantee that the TX datagram queue is empty if multiple threads are allowed to push the RESET or RESET-ACK packet.

Bug: 393354349
Test: CHQTS pass
Test: Run WWAN world, confirm link can be reset from both endpoints and continues functioning

Change-Id: I88aff7d2459af38d48a5c95ecfc0c6e49524fa8f