Initial code to avoid chreAbort on test failures
This is an initial CL which updates the basic WiFi and sensor flush async CHQTS test to avoid using chreAbort on test failures.

To avoid chreAbort, we replaced sendFatalFailureToHost calls with an EXPECT macro, which sends the failure to the host and then returns from the current method. The ASSERT_* macros are converted to EXPECT_* equivalents, to accurately portray to modified behavior, where the test continues to run, but the failure message will cause the nanoapp to be unloaded. Additional messages to the host (e.g. success, etc.) are rejected once the test has failed.

Bug: 304529053
Test: Confirm test nanoapp does not abort on assertion failures.
Test: Confirm multiple EXPECT_FAIL can happen without causing problems with the test reporting
Test: Confirm EXPECT_FAIL can be used right before sendSuccessToHost, and success message is rejected

Change-Id: I0151f2170f2263f62155539e10a1cc1793d3c1fd