Add IntrusiveRefBase
This is a base class for reference counting that is compatible
with pw::IntrusivePtr, which has a different API than the
existing RefBase. This provides similar functionality to
the Pigweed code, and is used instead to provide support for
CHRE platforms that do not support <atomic>, but do support
CHRE atomic abstractions.

This CL also adds some missing dependencies for pw_intrusive_ptr.

Bug: 373417024
Change-Id: Idb79d575f9b6627d14ef23538c596270790c6280
Flag: EXEMPT CHRE change
Test: Run chre_unit_tests, chre_simulation_tests