Triggers loopback when WiFi scan events are delayed
In some cases, WiFi scan events may be delayed for unexpected issues in the remote CHPP WiFi service. To handle this scenario, we add logic to schedule a loopback request when a WiFi scan event does not complete for a specific timeout.

To support this functionality, this CL adds the general capability for CHPP endpoints to schedule a timeout event, which is triggered from the transport work thread.

Bug: 390436314
Change-Id: If5aff854762ab81a087fbba90e7f15e8590cf00b
Test: Let WiFi world run for an extended period of time, confirm loopback triggers in problematic scenario, and the WiFi scan event + loopback result can be received
Test: atest chre_chpp_fake_link_sync_tests pass