Reset scan event accumulation logic on health monitor timeout
In a scenario where a WiFi scan event is delayed, it's possible that a scan request health monitor timeout can cause interference with the next scan request, because we remove the scan request in mPendingScanRequests, but don't reset the scan event accumulation logic. If a scan event arrives after the next scan request is made, the new request may be removed from mPendingScanRequests based on the scan event accumulation logic, if the delayed scan event completes the accumulator.

Bug: 399705346
Test: Confirm on device that we don't see a scan response from a request after a health monitor timeout to be dropped, which would otherwise be dropped without this change
Change-Id: I8aee82afcf98f590c5aaeaf6e701352e5ab2b389