Add wait functions for simulation tests
This CL adds built-in waiting functions for simulation tests.
This can be used where waitForEvent is not preferred - i.e.
when the ordering of events cannot be guaranteed or for more
simple code.

This CL also updates ChreMessageHubTest to use these new helpers and
deflakes some tests after making this change.

Bug: 397980248
Flag: TEST_ONLY
Test: Run chre_simulation_tests
Change-Id: I679537d66f63d2ca0aa557fa3800683a95b65e54